requirements.txt
APScheduler==3.11.2
google-api-python-client==2.193.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
anthropic
google-genai
python-dotenv==1.2.2
slack-bolt==1.27.0
fastapi
uvicorn[standard]
cryptography
